MorphoSSL/requirements.txt
torch
torchvision
timm
pytorch-lightning
einops
matplotlib
pillow
